Move hasMinimumStoryPosterDimensions to helpers as well
It’s not actually a UI component, just a boolean callback, which is better suited here.
