+++ assets/src/components/index.js
@@ -7,7 +7,6 @@ export { default as Shortcuts } from './shortcuts';
 export { default as StoryBlockDropZone } from './story-block-drop-zone';
 export { default as StoryBlockMover } from './block-mover';
 export { default as FontFamilyPicker } from './font-family-picker';
-export { default as hasMinimumStoryPosterDimensions } from './has-minimum-story-poster-dimensions';
 export { default as getFeaturedImageNotice } from './get-featured-image-notice';
 export { default as getPrePublishNotice } from './get-pre-publish-notice';
 export { default as withAmpStorySettings } from './with-amp-story-settings';
